Generating bipartite matched pairs

Generates matched pairs either:

With Replacement

Finds smallest control for each treatment

Without Replacement, Greedy

Greedily generates pairs. Note that the order for choosing the greedy pairs is random, which is not the only possible solution.

Without Replacement, Optimally

Minimum total distance

If you're happy to use control units potentially multiple times, then the first way is fast and optimal.

bipartite_matches(
  dist_mat,
  treat_vec,
  match_method = c("with_replacement", "optimal", "greedy"),
  n_sinks = NULL,
  tol_val = NULL,
  weight_vec = NULL
)

Arguments

dist_mat
Matrix of pairwise distances.

treat_vec
Vector representing all subjects; 0 for control, 1 for treated.

match_method
This enum corresponds to the three matching methods discussed above:

"with_replacement"

Finds smallest control for each treatment

"greedy"

Greedily generates pairs. Note that the order for choosing the greedy pairs is random, which is not the only possible solution.

"optimal"

Minimum total distance

n_sinks
how many sinks to use; can be vector. Note that for greedy and simple with-replacement matching, it's often better to sort this elsewhere. Optimal matching can only take one value. Default NULL to match all and ignore sinks.

tol_val
tolerance for solving optimal matches - how far is is acceptable to be from the true optimal value? Speed with large value, accuracy with small. Only relevant for !with_replacement && !greedy. Default 1e-4 is reasonable in many cases.

weight_vec
Default NULL: optionally supply the weight vector used to generate dist_mat and it'll be returned in the match_list generated from this function

Value

basic return value is a list with five elements and an optional sixth:

treat_index

index of treated units, from all units

treat_index_within

index of treated units, from the set of treated

control_index

index of control units, from all units

control_index_within

index of control units, from the set of control

distance

distances between the pairs

weight_vec

weight vector used to generate dist_mat if supplied

You'll get a list of such objects, each with an extra element: the number of sinks used. If you have n_sinks as NULL, then it'll default to a single sink value: zero.

Details

If not, you have to trade off speed vs optimality. Greedy runs over all units in a random order, so if you want to run greedy a bunch of times and take the best, it would still be (likely) much faster than running optimal matching.

Author

Colman Humphrey