Computes all matches, then gets the brier scores for each. Reorder by number of sinks.

Computes all matches, then gets the brier scores for each. Reorder by number of sinks.

brier_nonbipartite_matches(
  x_mat,
  cov_x,
  weight_list,
  tolerance_list = gen_tolerance_list(),
  match_method = c("with_replacement", "optimal", "greedy"),
  n_sinks = 0L,
  caliper_list = gen_caliper_list(),
  propensity_list = match_propensity_list(NULL),
  sqrt_mahal = TRUE,
  keep_all_with_replacement = FALSE,
  design = "cross_all",
  num_folds = 5,
  match_predict_function = match_predict_xgb(),
  silent = !interactive()
)

Arguments

x_mat
input/design matrix (already rank-adjusted etc)

cov_x
the (potentially rank-adjusted) covariance of x_mat. This means it's possible that cov(x_mat) is not equal to cov_x; see covariance_with_ranks for more details.

weight_list
list of weight vectors. See generate_random_weights to automatically generate a reasonable set of vectors.

tolerance_list
See gen_tolerance_list

match_method
This enum corresponds to the three matching methods discussed above:

"with_replacement"

Finds smallest control for each treatment

"greedy"

Greedily generates pairs. Note that the order for choosing the greedy pairs is random, which is not the only possible solution.

"optimal"

Minimum total distance

n_sinks
how many potential matches to not bother with NOTE: you can do this as a vector, but not for optimal matching.

caliper_list
Optional, see gen_caliper_list. Provide this to force matches that are close on some metric.

keep_all_with_replacement
logical, default FALSE. When nbp matching with replacement, you can in some cases form nearly as many pairs as there are units (if using a tolerance vec, then maybe not the lowest value pairs, since they'll have no control unit). But this would form a contrast with all other methods, so we will cut down to using half by default.

design
see predict_prepare

num_folds
how many CV folds to use

match_predict_function
function to predict treated units

Value

List of matches within sink values, and brier scores for each.

Author

Colman Humphrey