Abadie, Alberto and Guido W Imbens (2006). “Large sample properties of matching estimators for average treatment effects”. Econometrica 74.1, pp. 235–267. - see page 251. More info at “Matching on the estimated propensity score”. Econometrica 84.2, pp. 781–807.

Abadie, Alberto and Guido W Imbens (2006). “Large sample properties of matching estimators for average treatment effects”. Econometrica 74.1, pp. 235–267. - see page 251. More info at “Matching on the estimated propensity score”. Econometrica 84.2, pp. 781–807.

gen_bipartite_repeated_variance(
  x_mat,
  cov_x,
  y_vector,
  control_index,
  treat_vec,
  caliper_list = gen_caliper_list(),
  weight_vec = NULL,
  use_all_controls = TRUE,
  sqrt_mahal = TRUE
)

Arguments

x_mat
input/design matrix (already rank-adjusted etc)

cov_x
the (potentially rank-adjusted) covariance of x_mat. This means it's possible that cov(x_mat) is not equal to cov_x; see covariance_with_ranks for more details.

y_vector
Outcome vector (not used in match generation).

treat_vec
Logical (or 1/0) vector, indicating treatment (or control).

caliper_list
Optional, see gen_caliper_list. Provide this to force matches that are close on some metric.

use_all_controls
logical; if using bipartite matches, should we estimate the variance using all controls possible? Default TRUE. If FALSE, will only use controls from the match given