Solving GLS when variance is known up to a scale factor

If we know that \(y \sim \mathcal{N}(X \beta, \sigma^2 \Sigma)\) where \(\Sigma\) is known but \(\beta\) and \(\sigma^2\) aren't, then we can solve using nearly straight-forward generalised least squares. We return a list of results: the coefficients (\(\beta\)); the standard error for those coefficients; the whole covariance matrix for \(\beta\) and finally our estimate for \(\sigma^2\).

gls_known_relative_variance(y_vector, x_mat, rel_var_mat)

Arguments

y_vector
Outcome vector (in future can potentially be matrix)

x_mat
Matrix of explanatory variables. We will not include the intercept for you so add a column of 1s if you want this. You can also supply just a single vector and we'll upgrade it to a matrix (but again we will not add an intercept). Needs to match length of y_vector

rel_var_mat
\(\Sigma\) in the equations above: the variance of the error up to some (unknown) constant factor.

Value

List with the following elements:

beta_gls: the coefficients, \(\hat{\beta}\)

beta_gls_stderr: the standard errors for the coefficients

var_beta_gls: the covariance matrix for \(\beta\)

sigma_sq_est: our estimate for \(\sigma^2\). If x_mat has column names, they should hopefully pass through to the resulting vectors/matrices.

Author

Colman Humphrey