Simlar to bipartite_match_sd, except for use with non-bipartite matches.

For when your result isn't just based on the differences in the y-vector, but those differences are weighed inversely proportional to the difference between the tolerance values. To get on the standard error level, divide the result by sqrt(length(match_list[["treat_index"]])) This will be conservative in general with repeats.

nonbipartite_match_sd_scaled(
  x_mat,
  cov_x,
  y_vector,
  match_list,
  tolerance_list = gen_tolerance_list(),
  caliper_list = gen_caliper_list(),
  weight_vec = NULL,
  use_regression = TRUE,
  sqrt_mahal = TRUE
)

Arguments

x_mat
input/design matrix (already rank-adjusted etc)

cov_x
the (potentially rank-adjusted) covariance of x_mat. This means it's possible that cov(x_mat) is not equal to cov_x; see covariance_with_ranks for more details.

y_vector
Outcome vector (not used in match generation).

match_list
A particular match result.

tolerance_list
See gen_tolerance_list

caliper_list
Optional, see gen_caliper_list. Provide this to force matches that are close on some metric.

weight_vec
Default NULL: optionally supply the weight vector used to generate dist_mat and it'll be returned in the match_list generated from this function

Value

a single float, the standard deviation (not standard error) for the match

Author

Colman Humphrey