Generating nonbipartite matched pairs

Generates matched pairs either:

With Replacement

Finds smallest control for each treatment

Without Replacement, Greedy

Greedily generates pairs. Note that the order for choosing the greedy pairs is random, which is not the only possible solution.

Without Replacement, Optimally

Minimum total distance

If you're happy to allow units to be the control multiple times, then the first way is fast and optimal. If not, you have to trade off speed vs optimality. Greedy runs over all units in a random order, so if you want to run greedy a bunch of times and take the best, it would still be (likely) much faster than running optimal matching.

nonbipartite_matches(
  dist_mat,
  tolerance_list = gen_tolerance_list(),
  match_method = c("with_replacement", "optimal", "greedy"),
  n_sinks = NULL,
  keep_all_with_replacement = FALSE
)

Arguments

dist_mat
Matrix of pairwise distances.

tolerance_list
See gen_tolerance_list

match_method
This enum corresponds to the three matching methods discussed above:

"with_replacement"

Finds smallest control for each treatment

"greedy"

Greedily generates pairs. Note that the order for choosing the greedy pairs is random, which is not the only possible solution.

"optimal"

Minimum total distance

n_sinks
Vector of sinks per match. Note that this is NOT the same for each match_method: for both "with_replacement" and "greedy", it subtracts one pair. But for optimal matching, it removes one full unit. "greedy" is the weird one here, but it wouldn't truly make sense given the code to try and replicate the optimal implementation. Be safe and don't use greedy...

keep_all_with_replacement
logical, default FALSE. When nbp matching with replacement, you can in some cases form nearly as many pairs as there are units (if using a tolerance vec, then maybe not the lowest value pairs, since they'll have no control unit). But this would form a contrast with all other methods, so we will cut down to using half by default.

Value

basic return value is a list with three elements:

treat_index

index of treated units

control_index

index of control units

distance

distances between the pairs

If n_sinks is not NULL, you'll get a list of such objects, each with an extra element: the number of sinks used.

Details

NOTE: with sinks - the optimal match gives "half" the number of sinks you might expect. If you have a pool of 400 matchable things, NBP gives you 200 matches if no sinks. If you say sinks = 20 FOR OPTIMAL, you'll get 190 matches, not 180. We leave this behaviour in so that this algorithm can generate expected results if someone is used to nbpmatch Further, note that we don't allow equality on the tolerance vector if used.