Given a vector of sink values, produces optimal NBP matches for each

Given a vector of sink values, produces optimal NBP matches for each

optimal_nbp_sink_wrap(
  dist_mat,
  tolerance_vec,
  n_sinks = NULL,
  weight_vec = NULL
)

Arguments

dist_mat
Matrix of pairwise distances.

n_sinks
Vector of sinks per match. Note that this is NOT the same for each match_method: for both "with_replacement" and "greedy", it subtracts one pair. But for optimal matching, it removes one full unit. "greedy" is the weird one here, but it wouldn't truly make sense given the code to try and replicate the optimal implementation. Be safe and don't use greedy...

weight_vec
Default NULL: optionally supply the weight vector used to generate dist_mat and it'll be returned in the match_list generated from this function

Author

Colman Humphrey