Takes matches and their Brier scores, and computes permutation Brier scores and the best matches

Can work for bipartite or non-bipartite matches, the permutation is just over the labels.

Usage

permutation_matches(
  matches_by_sinks,
  briers_by_sinks,
  x_mat,
  n_sinks = 0L,
  approximate_by_best = TRUE,
  silent = !interactive()
)

Arguments

matches_by_sinks

List by number of sinks, each a list of match results (a match list), for each weight vector.

briers_by_sinks

List by number of sinks, each a vector of Brier results. Basically a number for each match in matches_by_sinks.

x_mat

Typical input matrix

n_sinks

Vector of number of sinks - probably could get this directly from matches_by_sinks, but nice to be explicit.

approximate_by_best

Logical, default TRUE. Only compute one permutation distribution, using the best result by brier score to do so. Useful because it changes little, but saves a ton of time.

silent

Do you want to suppress message output? Default !interactive().

Value

Returns a list of two lists. The first is vectors of permutation Brier scores (one per match). The second is the best match at each sink value, along with some extra info about that match.

Author

Colman Humphrey