Prepares design matrices and other inputs needed for predictive methods

It is not immediately clear how to take a match (set of matched pairs) and "predict" it - essentially you likely want some prediction function \(f(x, y) -> {0, 1}\) (or \(P(1)\) etc) where \(x\) and \(y\) are rows, one of which is treated and the other control. This function should also have a property resembling: $$f(x, y) = 1 - f(y, x)$$ since there's an inherent symmetry in the pairs: if you knew which of the two came first, there'd be nothing to do. In fact, the defition of \(f\) would likely be to predict if (WLOG) \(x\) is the treated unit or not.

predict_prepare(
  x_mat,
  index_list,
  design = c("cross_all", "cross_random", "differences_random", "differences_plain")
)

Arguments

x_mat
original data, i.e. row for each unit

index_list
list of indices to use in the matching:

treat_train

Index of treated units for the training matrix

control_train

Index of control units for the training matrix

treat_test

Index of treated units for the test matrix

control_test

Index of control units for the test matrix

design
The design matrix form do you want to use. We'll assume that x_mat is \(n \times k\). Let \(X_c\) be the rows of x_mat corresponding to the control units, and \(X_t\) the treated units that pair with \(X_c\) (i.e. so the \(i^{\text{th}}\) matched pair will correspond with the \(i^{\text{th}}\) row of each), and thus each matrix will have dimension \(m \times k\), where \(2m \leq n\).

"cross_all"

This will form the \(2m \times 2k\) matrix, by having each matched pair give two rows in the design matrix: each of the two rows will contain the full treated row of the given pair along with the full control row, but in one row the pair will be (treated, control), and in the other (control, treated). For these two rows, the output vector will contain a 1 and a 0 respectively. We construct the following inputs: $$X = \begin{bmatrix} X_t& X_c \\ X_c& X_t \\ \end{bmatrix}, Y = [1 1 \hdots 1 0 0 \hdots 0]' $$ For the test data, we'll only (WLOG more or less) give the (treat, control) version, for consistency in evaluation.

"cross_random"

This is similar to the above, except instead of forming all rows in both ways, we choose (for each pair) between (treated, control) and (control, treated) randomly. So $$Y_i = \begin{cases} 1 & \text{with probability} 0.5 \\ 0 & \text{with probability} 0.5 \end{cases}$$ And (\(X_i\) here meaning the \(i^{\text{th}}\) row of the design matrix): $$X_i = \begin{cases} ((X_t)_i, (X_c)_i) & Y_i = 1 \\ ((X_c)_i, (X_t)_i) & Y_i = 0 \end{cases}$$ Giving an \(m \times 2k\) design matrix.

"differences_random"

For each pair, our design row will be the row in x_mat corresponding to one unit minus the row corresponding to the other. Similar to the above, we'll choose the outcome randomly and decide which unit will be subtracted from the other based on that, and we'll end up with an \(m \times k\) matrix. That is: $$Y_i = \begin{cases} 1 & \text{with probability} 0.5 \\ 0 & \text{with probability} 0.5 \end{cases}$$ And: $$X_i = \begin{cases} (X_t)_i - (X_c)_i & Y_i = 1 \\ (X_c)_i - (X_t)_i & Y_i = 0 \end{cases}$$ Note that if your x_mat contains an intercept, it'll now be a column of zeroes.

"differences_plain"

In some cases, you may wish to have the differences and add customization. So here we essentially do the same as above but with \(Y_i = 1\) for all \(i\), therefore all differences are returned as \((X_t)_i - (X_c)_i\) (and will also be \(m \times k\)).

Value

list:

x_train

Design matrix for training

x_test

Design matrix for testing

y_train

Training outcome vector

y_test

Test outcome vector

Details

This function gives various potential design matrix shapes and corresponding output vectors that all can be used to solve the above prediction problem. In some scenarios (e.g. if we were content to use linear models), differences alone might be appropriate. For full generality, we may wish to use both full vectors.