Computes weighted Mahalanobis distance, using Choleski decomp.

Note: def. of weighted: \(d(x_i, x_j) = (x_i - x_j)' W \Sigma^{-1} W (x_i - x_j)\) where \(W\) = diag(weight_vec) R's cholesky gives \(U\) s.t. \(U' U = S\) (i.e. \(U = chol(S)\)) so in general we want: $$ (x_i - x_j)' W (U' U)^{-1} W (x_i - x_j) \\ = (x_i - x_j)' W U^{-1} (U')^{-1} W (x_i - x_j) \\ = x_i' W U^{-1} (U')^{-1} W x_i + x_j' W U^{-1} (U')^{-1} W x_j \\ - 2 x_i' W U^{-1} (U')^{-1} W x_j $$ Solving the above is manageable if we have \(y_i = (U')^{-1} W x_i\) or moving terms around, \(W^{-1} U' y_i = x_i\), which is simple by forwardsolve. Then we have: $$dist(x_i, x_j) = ||y_i||^2 + ||y_j||^2 - 2 y_i' y_j$$ Letting \(Y = (y_1', y_2', .... ,y_n')'\) (which we can get in one line, \(Y' = \) forwardsolve \((W^{-1} U', x_mat')\)) the first two parts are just rowSums\((Y^2)\) (note that the code uses \(Y'\), thus colSums) (add outer(.,.) to finish) and the last is \(Y Y'\)

Usage

weighted_mahal(
  x_mat,
  cov_x,
  weight_vec = NULL,
  treat_vec = NULL,
  sqrt_mahal = TRUE,
  partial_index = NULL
)

Arguments

x_mat

numeric matrix (adjust non-numeric columns prior), already rank-adjusted if desired

cov_x

covariance of x, calculated potentially with ranks

weight_vec

vector of weights corresponding to columns of x_mat, giving weights relative to "raw" (ranked) Mahalanobis. Note that the resulting matrix does depend on the scale of the weight vector, but the matching won't: scaling the Mahalanobis matrix has no effect on distance minimising pairs etc

treat_vec

optionally specify which units are treated. If NULL (default), will just return nrow(x_mat) x nrow(x_mat) distance matrix of all pairs. Can be logicals or \({0, 1}\)

sqrt_mahal

logical, default TRUE; do you want regular Mahalanobis: \(d(x_i, x_j) = (x_i - x_j)' \Sigma^{-1} (x_i - x_j)\) or the square root? (in weighted, \(\Sigma^{-1}\) becomes \(W \Sigma^{-1} W\) in the above)

partial_index

In some cases, you want a subset of the full N x N matrix, but you can't partition it like you want with treat_vec. e.g. you want full_dist[c(1, 2, 3), 1:10] then use partial_index = list(c(1,2,3), 1:10)

Value

returns a matrix of pairwise distances; the relevant indexing depends on treat_vec and partial_index